// src/app/modules/room/janus/janus.interfaces.d.ts
declare namespace JanusJS {
// ...
  interface JSEP {}
// ...
  interface Publisher {
    id: number;
    display: string;
    streams: any[];
    audio_codec?: string;
    video_codec?: string;
    talking?: boolean;
  }

  enum MessageType {
    Recording = 'recording',
    Starting = 'starting',
    Started = 'started',
    Stopped = 'stopped',
    SlowLink = 'slow_link',
    Preparing = 'preparing',
    Refreshing = 'refreshing'
  }

  interface Message {
    result?: {
      status: MessageType;
      id?: string;
      uplink?: number;
    };
    error?: Error;
    publishers?: Publisher[];
  }
// ...
  interface PluginMessage {
    message: {
      request: string;
      [otherProps: string]: any;
    };
    jsep?: JSEP;
    success?: Function;
    error?: (error: any) => void;
  }

  interface PluginHandle {
    plugin: string;
    id: string;
    token?: string;
    detached : boolean;
    webrtcStuff: {
      started: boolean,
      myStream: MediaStream,
      streamExternal: boolean,
      remoteStream: MediaStream,
      mySdp: any,
      mediaConstraints: any,
      pc: RTCPeerConnection,
      dataChannelOptions: RTCDataChannelInit,
      dataChannel: Array<RTCDataChannel>,
      dtmfSender: any,
      trickle: boolean,
      iceDone: boolean,
      volume: {
        value: number,
        timer: number
      }
    };
    getId(): string;
    getPlugin(): string;
    send(message: PluginMessage): void;
    createOffer(params: any): void;
    createAnswer(params: any): void;
    handleRemoteJsep(params: { jsep: JSEP }): void;
    dtmf(params: any): void;
    data(params: any): void;
    isAudioMuted(): boolean;
    muteAudio(): void;
    unmuteAudio(): void;
    isVideoMuted(): boolean;
    muteVideo(): void;
    unmuteVideo(): void;
    getBitrate(): string;
    hangup(sendRequest?: boolean): void;
    detach(params: any): void;
  }
// ...
}